converted to composable to streamFieldUtils
